+++ changelog.md
@@ -1,3 +1,4 @@
 v1.0.1
 - Fix splashscreen getting stuck
 - Fix subscribing to thread causing a crazy unread posts count
+- Fix json parsing for notifications #61